feat(form/gerar): Extrai lógica para hook e exibe link de acesso ao formulário
